JS File
dd a listener for click events on the save button.
Add code to apply the past, present, or future class to each time block by comparing the id to the current hour.

Add code to get any user input that was saved in localStorage and set the values of the corresponding textarea elements